Update data-counter
Thank you for submitting this.

For the `data-counter` syntax, I notice that the yaml syntax with true wasn't a good idea. In fact, the js only checks for the presence of the attribute regardless of the value.
So even `data-counter: false` would work. I've just found that `~` means null in markdown, so it's probably a better choice.